+++ README.md
@@ -109,7 +109,8 @@ BloxyApi api = new BloxyApi("YourApiKey", supplier);
 
 *Kotlin*
 ```kotlin
-val api = BloxyApi("YourApiKey", Supplier { HttpClient(6000, 10000) })
+val supplier = Supplier { HttpClient(10000, 40000) }
+val api = BloxyApi("YourApiKey", supplier)
 ```
 
 ## API Examples
