--- src/main/kotlin/io/api/bloxy/error/ParamException.kt
@@ -1,10 +1,0 @@
-package io.api.bloxy.error
-
-
-/**
- * Invalid API parameter error
- *
- * @author GoodforGod
- * @since 21.11.2018
- */
-class ParamException(message: String?) : BloxyException(message)
